refactor: extract logic by platforms using interface
